project/requirements.txt
dlib==19.24.0
imageio==2.23.0
matplotlib==3.6.2
numpy==1.24.1
opencv_python==4.7.0.68
pandas==1.5.2
Pillow==9.4.0
scipy==1.10.0
